+++ cvopt/model_selection/__init__.py
@@ -1,4 +1,4 @@
 # coding: utf-8
-from ._search import SimpleoptCV, HyperoptCV, BayesoptCV, GAoptCV
+from ._search import SimpleoptCV, RandomoptCV, HyperoptCV, BayesoptCV, GAoptCV
 
-__all__ = ("SimpleoptCV", "HyperoptCV", "BayesoptCV", "GAoptCV")
+__all__ = ("SimpleoptCV", "RandomoptCV", "HyperoptCV", "BayesoptCV", "GAoptCV")