+++ cvopt/model_selection/_search.py
@@ -31,7 +31,7 @@ class SimpleoptCV():
     cv: scikit-learn cross-validator or int(number of folds), default=5.
         Cross validation setting.
 
-    max_iter: int, default=10.
+    max_iter: int, default=32.
         Number of search.
 
     random_state: int or None, default=None.
@@ -91,7 +91,15 @@ class SimpleoptCV():
         Refit an estimator using the best found parameters on all train data(=X).
 
     backend: str, default="hyperopt".
-        backend optimeizer.
+        backend optimeizer. Supports the following back ends.
+
+        * `hyperopt`: Sequential Model Based Global Optimization
+
+        * `bayesopt`: Bayesian Optimization
+
+        * `gaopt`: Genetic Algorithm
+
+        * `randomopt`: Random Search
 
     Attributes
     ----------
@@ -109,7 +117,7 @@ class SimpleoptCV():
         Parameter setting that gave the best results on the hold out data.
     """      
     def __init__(self, estimator, param_distributions, 
-                 scoring=None, cv=5, max_iter=10, 
+                 scoring=None, cv=5, max_iter=32, 
                  random_state=None, n_jobs=1, pre_dispatch="2*n_jobs", 
                  verbose=0, logdir=None, save_estimator=0, model_id=None, refit=True, 
                  backend="hyperopt", **kwargs): 
@@ -131,8 +139,14 @@ class SimpleoptCV():
                                  n_jobs=n_jobs, pre_dispatch=pre_dispatch, verbose=verbose, logdir=logdir, 
                                  save_estimator=save_estimator, model_id=model_id, refit=refit, 
                                  **kwargs)
+        elif backend == "randomopt":
+            self.optcv = RandomoptCV(estimator, param_distributions, 
+                                     scoring=scoring, cv=cv, max_iter=max_iter, random_state=random_state, 
+                                     n_jobs=n_jobs, pre_dispatch=pre_dispatch, verbose=verbose, logdir=logdir, 
+                                     save_estimator=save_estimator, model_id=model_id, refit=refit, 
+                                     **kwargs)
         else:
-            raise Exception("`backend` is must be `hyperopt` or `bayesopt`.")
+            raise Exception("`backend` "+str(backend)+" is not supported.")
 
         self.backend = backend
             
@@ -160,7 +174,7 @@ class HyperoptCV(BaseSearcher):
     cv: scikit-learn cross-validator or int(number of folds), default=5.
         Cross validation setting.
 
-    max_iter: int, default=10.
+    max_iter: int, default=32.
         Number of search.
 
     random_state: int or None, default=None.
@@ -238,7 +252,7 @@ class HyperoptCV(BaseSearcher):
         Parameter setting that gave the best results on the hold out data.
     """
     def __init__(self, estimator, param_distributions, 
-                 scoring=None, cv=5, max_iter=10, 
+                 scoring=None, cv=5, max_iter=32, 
                  random_state=None, n_jobs=1, pre_dispatch="2*n_jobs", 
                  verbose=0, logdir=None, save_estimator=0, model_id=None, refit=True, 
                  algo=tpe.suggest):
@@ -323,7 +337,7 @@ class BayesoptCV(BaseSearcher):
     cv: scikit-learn cross-validator or int(number of folds), default=5.
         Cross validation setting.
 
-    max_iter: int, default=10.
+    max_iter: int, default=32.
         Number of search.
 
     random_state: int or None, default=None.
@@ -480,7 +494,7 @@ class BayesoptCV(BaseSearcher):
     """
 
     def __init__(self, estimator, param_distributions, 
-                 scoring=None, cv=5, max_iter=10, 
+                 scoring=None, cv=5, max_iter=32, 
                  random_state=None, n_jobs=1, pre_dispatch="2*n_jobs", 
                  verbose=0, logdir=None, save_estimator=0, model_id=None, refit=True, 
                  max_time=np.inf, model_type="GP", initial_params=None, initial_score=None, 
@@ -596,7 +610,7 @@ class GAoptCV(BaseSearcher):
     cv: scikit-learn cross-validator or int(number of folds), default=5.
         Cross validation setting.
 
-    max_iter: int, default=10.
+    max_iter: int, default=32.
         Number of search.
 
     random_state: int or None, default=None.
@@ -655,11 +669,13 @@ class GAoptCV(BaseSearcher):
     refit: bool, default=True.
         Refit an estimator using the best found parameters on all train data(=X).
 
-    iter_pergeneration: int, default=5.
+    iter_pergeneration: int, default=8.
         Genetic algorithm's parameter. Number of iteration per generation (it corresponds to number of population.).
 
     param_crossover_proba: float or function, default=0.5.
         Genetic algorithm's parameter. Probability which a certain parameter becomes another parent value.
+
+        If this value 0 or 1, paramaters is not changed by crossover.
 
         Function whose variable is number of generation could be passed to this variable.
         Number of generation' s start is 0. But create population by random sampling in generation 0, so this function is used from generation 1.
@@ -695,10 +711,10 @@ class GAoptCV(BaseSearcher):
         Parameter setting that gave the best results on the hold out data.
     """
     def __init__(self, estimator, param_distributions, 
-                 scoring=None, cv=5, max_iter=10, 
+                 scoring=None, cv=5, max_iter=32, 
                  random_state=None, n_jobs=1, pre_dispatch="2*n_jobs", 
                  verbose=0, logdir=None, save_estimator=0, model_id=None, refit=True, 
-                 iter_pergeneration=5, param_crossover_proba=0.5, param_mutation_proba=0.01, 
+                 iter_pergeneration=8, param_crossover_proba=0.5, param_mutation_proba=0.01, 
                  random_sampling_proba=0.01):
         super().__init__(estimator=estimator, param_distributions=param_distributions, 
                          scoring=scoring, cv=cv,  n_jobs=n_jobs, pre_dispatch=pre_dispatch, 
@@ -759,6 +775,167 @@ class GAoptCV(BaseSearcher):
             gamin(obj, param_distributions, max_iter=self.max_iter, iter_pergeneration=self.iter_pergeneration, 
                   param_crossover_proba=self.param_crossover_proba, param_mutation_proba=self.param_mutation_proba, 
                   random_sampling_proba=self.random_sampling_proba, cvsummarizer=self._cvs)
+        except KeyboardInterrupt:
+            pass
+
+        self._postproc_fit(X=X, y=y, feature_groups=feature_groups, 
+                           best_params=self._cvs.best_params_, best_score=self._cvs.best_score_)
+        return self
+
+
+
+class RandomoptCV(BaseSearcher):
+    """
+    Cross validation optimizer by Random Search.
+
+    Parameters
+    ----------
+    estimator
+        scikit-learn estimator like.
+
+    param_distributions: dict.
+        Search space.
+
+    scoring: string or sklearn.metrics.make_scorer.
+        Evaluation index of search.
+        When scoring is None, use stimator default scorer and this score greater is better.
+        
+    cv: scikit-learn cross-validator or int(number of folds), default=5.
+        Cross validation setting.
+
+    max_iter: int, default=32.
+        Number of search.
+
+    random_state: int or None, default=None.
+        The seed used by the random number generator.
+
+    n_jobs: int, default=1.
+        Number of jobs to run in parallel.
+
+    pre_dispatch: int or string, default="2*n_jobs".
+        Controls the number of jobs that get dispatched during parallel.
+
+    verbose: int(0, 1 or 2), default=0.
+        Controls the verbosity
+        
+        0: don't display status.
+
+        1: display status by stdout.
+        
+        2: display status by graph.
+
+    logdir: str or None, default=None.
+        Path of directory to save log file.
+        When logdir is None,  log is not saved.
+        
+        [directory structure]
+        
+        logdir
+        
+        |-cv_results
+        
+        | |-{model_id}.csv                                      : search log
+        
+        | ...
+        
+        |-estimators_{model_id}
+        
+            |-{model_id}_index{search count}_split{fold count}.pkl: an estimator which is fitted fold train data
+            
+            ...
+            
+            |-{model_id}_index{search count}_test.pkl             : an estimator which is fitted whole train data.
+
+    save_estimator: int, default=0.
+        estimator save setting.
+        
+        0: An estimator is not saved.
+        
+        1: An estimator which is fitted fold train data is saved per cv-fold.
+        
+        2: In addition to 1, an estimator which is fitted whole train data is saved per cv.
+
+    model_id: str or None, default=None.
+        This is used to log filename.
+        When model_id is None, this is generated by date time.
+
+    refit: bool, default=True.
+        Refit an estimator using the best found parameters on all train data(=X).
+
+    Attributes
+    ----------
+    cv_results_ : dict of numpy (masked) ndarrays
+        A dict with keys as column headers and values as columns, that can be
+        imported into a pandas ``DataFrame``.
+
+    best_estimator_ : estimator or dict
+        Estimator that was chosen by the search.
+
+    best_score_ : float
+        Cross-validated score of the best_estimator.
+
+    best_params_ : dict
+        Parameter setting that gave the best results on the hold out data.
+    """
+    def __init__(self, estimator, param_distributions, 
+                 scoring=None, cv=5, max_iter=32, 
+                 random_state=None, n_jobs=1, pre_dispatch="2*n_jobs", 
+                 verbose=0, logdir=None, save_estimator=0, model_id=None, refit=True):
+        super().__init__(estimator=estimator, param_distributions=param_distributions, 
+                         scoring=scoring, cv=cv,  n_jobs=n_jobs, pre_dispatch=pre_dispatch, 
+                         verbose=verbose, logdir=logdir, save_estimator=save_estimator, 
+                         model_id=model_id, refit=refit, backend="gaopt")
+
+        self.max_iter = max_iter
+        if random_state is None:
+            self.random_state = random_state
+        else:
+            self.random_state = np.random.RandomState(int(random_state))
+
+    def fit(self, X, y=None, validation_data=None, groups=None, 
+            feature_groups=None, min_n_features=2):
+        """
+        Run fit.
+
+        Parameters
+        ----------       
+        X :numpy.array, pandas.DataFrame or scipy.sparse, shape(axis=0) = (n_samples)
+            Features. Detail depends on estimator.
+
+        y: np.ndarray or pd.core.frame.DataFrame, shape(axis=0) = (n_samples) or None, default=None.
+            Target variable. detail depends on estimator.
+
+        validation_data: tuple(X, y) or None, default=None.
+            Data to compute validation score. detail depends on estimator.
+            When validation_data is None, computing validation score is not run.
+
+        groups: array-like, shape = (n_samples,)  or None, default=None.
+            Group labels for the samples used while splitting the dataset into train/test set.
+            (input of scikit-learn cross-validator)
+
+        feature_groups: array-like, shape = (n_samples,) or None, default=None.
+            Group labels for the features used while fearture select.
+            When feature_groups is None, fearture selection is not run.
+
+        min_n_features: int, default=2.
+            When number of X's feature cols is less than min_n_features, return search failure.
+            
+            e.g. If estimator has columns sampling function, use this option to avoid X become too small and error.
+        """
+        X, y, Xvalid, yvalid, cv, param_distributions = self._preproc_fit(X=X, y=y, validation_data=validation_data, feature_groups=feature_groups)
+        np.random.seed(self.random_state)
+
+        obj = mk_objfunc(X=X, y=y, groups=groups, feature_groups=feature_groups, feature_axis=BaseSearcher.feature_axis, 
+                         estimator=self.estimator, scoring=self.scoring, cv=cv, 
+                         param_distributions=param_distributions, backend=self.backend, failedscore=np.nan, 
+                         score_summarizer=BaseSearcher.score_summarizer, 
+                         Xvalid=Xvalid, yvalid=yvalid, n_jobs=self.n_jobs, pre_dispatch=self.pre_dispatch, 
+                         cvsummarizer=self._cvs, save_estimator=self.save_estimator, min_n_features=min_n_features)
+
+        try :
+            gamin(obj, param_distributions, max_iter=self.max_iter, iter_pergeneration=1, 
+                  param_crossover_proba=0, param_mutation_proba=0, 
+                  random_sampling_proba=1, cvsummarizer=self._cvs)
         except KeyboardInterrupt:
             pass